+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Validate Canadian postal code format
+- Enforce character limit for location name so it doesn't run off the poster
 - Add QR Alert docs to Portal
 - Add Location Outbreaks History view
 - New venue types and selection html select control
+- Content updates throughout
+
 
 ## [2.0.3] - 2021-05-14